refactor(test): extract create browser commands (#262)
